refactor(server): rename docker wrapper
